test: CartList 컴포넌트 스토리북 테스트 컨텍스트 적용 및 기본 장바구니 상태 적용
